+++ halotools/mock_observables/two_point_clustering/__init__.py
@@ -6,7 +6,9 @@ galaxy-galaxy lensing `~halotools.mock_observables.delta_sigma`, and more.
 """
 from __future__ import absolute_import 
 
-__all__ = ('angular_tpcf', 'delta_sigma')
+__all__ = ('angular_tpcf', 'delta_sigma', 's_mu_tpcf', 'tpcf_multipole')
 
 from .angular_tpcf import angular_tpcf 
 from .delta_sigma import delta_sigma
+from .s_mu_tpcf import s_mu_tpcf
+from .tpcf_multipole import tpcf_multipole
